Integrate MirrorConfig
